Don't treat quoted literals as process calls
In a shell, most things end up being a process call. If a statement
starts with something that isn't a keyword, it is usually a call. This
may not always be desirable when expressing code that don't deal with
external executables. Due to the fact that all expressions have a type,
the user might want to return a string from a condition, and would be
surprised that it turned into a call. When being in the scripting part
of the language, it becomes difficult to avoid accidentally invoking
commands. On the other hand, we don't want to deny the user from being
able to call an executable with a name determined at runtime. The
compromise implemented here is that if there are any quotes, the parser
treat that as a single literal, and not as a call. If the user still
wants to use this syntax, they can prefix the statement with the `dyn`
keyword. This explicit the fact that is some external stuff that should
not be checked as rigorously as a call would.