Create a Docker image with glibc+BusyBox
The Alpine version was never updated to 3.20 due to incompatibilities
with musl build+gcompat. Rust struggles to generate a musl binary
when musl-gcc is used[^1], and creates something that always segfaults
as of now. Because this image is not really a base image for anything
else, a package manager is not really necessary, so a simple BusyBox
image is sufficient.

[^1]: https://www.github.com/rust-lang/rust/issues/95926